Move *Flags out of mod fuse_abi
For consistency, either all flags should be inside `fuse_abi`, or outside.
Previously I put some flag definifions outside, and this PR moves earlier
created flags outside too.

Personally I prefer smaller modules, but we can do the other way around:
move other flags into `mod fuse_abi`.
