+++ src/reply.rs
@@ -20,11 +20,11 @@ use log::warn;
 use crate::Errno;
 use crate::FileAttr;
 use crate::FileType;
+use crate::FopenFlags;
 use crate::PollEvents;
 use crate::channel::ChannelSender;
 use crate::ll::Generation;
 use crate::ll::INodeNo;
-use crate::ll::fuse_abi::FopenFlags;
 use crate::ll::reply::DirEntList;
 use crate::ll::reply::DirEntOffset;
 use crate::ll::reply::DirEntPlusList;
